% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar_method.R
\name{.DollarNames.treelabel}
\alias{.DollarNames.treelabel}
\alias{$.treelabel}
\title{Dollar access the underlying values inside a treelabel}
\usage{
\method{.DollarNames}{treelabel}(x, pattern = "")

\method{$}{treelabel}(x, name)
}
\arguments{
\item{x}{the treelabel vector}

\item{pattern}{the search pattern that is matched against the vertices
in the tree.}

\item{name}{the name of the vertex in the tree}
}
\value{
a boolean or numeric vectors with the column from \code{tl_score_matrix(x)[,name]}
}
\description{
Dollar access the underlying values inside a treelabel
}
