% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_across.R
\name{tl_across}
\alias{tl_across}
\alias{tl_if_any}
\alias{tl_if_all}
\alias{tl_sum_across}
\alias{tl_mean_across}
\alias{tl_countNAs_across}
\title{dplyr-style across functions for treelabels}
\usage{
tl_across(.cols, expr, ...)

tl_if_any(.cols, expr, ...)

tl_if_all(.cols, expr, ...)

tl_sum_across(.cols, expr, ..., na.rm = TRUE)

tl_mean_across(.cols, expr, ..., na.rm = TRUE)

tl_countNAs_across(.cols, expr, ..., na.rm = TRUE)
}
\arguments{
\item{.cols}{tidyselect expression of the column which contain the treelabel
columns}

\item{expr}{an unquoted expression evaluated with \code{tl_eval} for each column.}

\item{...}{addtional parameters passed on to \code{across}}

\item{na.rm}{boolean that decides if missing values are ignored.}
}
\description{
dplyr-style across functions for treelabels
}
\seealso{
\code{\link[dplyr:across]{dplyr::across()}}
}
