% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_as_logical}
\alias{tl_as_logical}
\alias{tl_as_numeric}
\title{Convert the underlying data-representation from logical to numeric}
\usage{
tl_as_logical(x)

tl_as_numeric(x)
}
\arguments{
\item{x}{\code{treelabel} vector}
}
\description{
Convert the underlying data-representation from logical to numeric
}
