% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_atmost}
\alias{tl_atmost}
\title{Replace each NA with the maximum value that it could have}
\usage{
tl_atmost(x)
}
\arguments{
\item{x}{\code{treelabel} vector}
}
\value{
the modified \code{treelabel} vector
}
\description{
In pseudo-notation, this function calculates for any node \code{n} that is \code{NA},
\code{score = score(parent(n)) - sum(children(parent(n)), na.rm=TRUE)}.
}
