% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_eval.R
\name{tl_eval}
\alias{tl_eval}
\title{Evaluate an expression in the tree label environment}
\usage{
tl_eval(x, expr, ...)
}
\arguments{
\item{x}{`treelabel` vector}

\item{expr}{an unquoted expression that is evaluated in the treelabel
environment. Can, for example, be used to check if a particular
score is higher than a threshold}

\item{...}{additional arguments passed on to `rlang::eval_tidy`}
}
\description{
Evaluate an expression in the tree label environment
}
