% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_eval.R
\name{tl_get_score}
\alias{tl_get_score}
\title{Get the score for a name of a tree node}
\usage{
tl_get_score(x, name)
}
\arguments{
\item{x}{`treelabel` vector}

\item{name}{the name of one of the nodes in the tree}
}
\description{
Get the score for a name of a tree node
}
