% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_modify}
\alias{tl_modify}
\title{Modify the treelabel values}
\usage{
tl_modify(x, ..., .propagate_NAs_down = TRUE)
}
\arguments{
\item{x}{object that is converted to a \code{treelabel}}

\item{...}{tidyeval dots which work like \code{dplyr::mutate()}.}

\item{.propagate_NAs_down}{boolean that decides if the score of children is to \code{NA} if
their parents score is \code{NA}.}
}
\value{
the modified treelabel vector
}
\description{
Modify the treelabel values
}
