% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_modify_tree}
\alias{tl_modify_tree}
\title{Modify tree of an existing treelabel}
\usage{
tl_modify_tree(x, new_tree, tree_root = tl_tree_root(x), ...)
}
\arguments{
\item{x}{\code{treelabel} vector}

\item{new_tree}{an igraph object that describes the new tree}

\item{tree_root}{the name of the tree root. Defaults to the
to the name of the tree root of \code{x}.}

\item{...}{arguments passed to \code{\link{treelabel.matrix}}}
}
\value{
a \code{treelabel} vector described by the new tree.
}
\description{
Modify tree of an existing treelabel
}
