% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_name.R
\name{tl_name}
\alias{tl_name}
\title{Get a single name for a treelabel}
\usage{
tl_name(x, threshold = 0)
}
\arguments{
\item{x}{`treelabel` vector}

\item{threshold}{the threshold against which each score is compared.}
}
\description{
Returns the name of the first node from the bottom where the score
is higher than the threshold. If multiple nodes pass the threshold and are equally far away
from the root, the one with the higher score is returned.
}
