% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_replace_NAs}
\alias{tl_replace_NAs}
\title{Replace \code{NA}s with zeros}
\usage{
tl_replace_NAs(x)
}
\arguments{
\item{x}{\code{treelabel} vector}
}
\description{
Replaces all \code{NA}s except in score matrix with zeros, except for the
"root" column. This function is useful for comparing values.#'
}
