% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_score_matrix.R
\name{tl_score_matrix}
\alias{tl_score_matrix}
\title{Low-level function to access the data underlying a tree label}
\usage{
tl_score_matrix(x)
}
\description{
Low-level function to access the data underlying a tree label
}
