% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelabel.R
\name{tl_tree}
\alias{tl_tree}
\alias{tl_tree_root}
\title{Extract tree or root from treelabel}
\usage{
tl_tree(x)

tl_tree_root(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract tree or root from treelabel
}
