% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_set_score.R
\name{tl_tree_modify}
\alias{tl_tree_modify}
\alias{tl_tree_filter}
\alias{tl_tree_cut}
\title{Modify tree of an existing treelabel}
\usage{
tl_tree_modify(x, new_tree, tree_root = tl_tree_root(x), ...)

tl_tree_filter(x, .p, ...)

tl_tree_cut(x, new_root)
}
\arguments{
\item{x}{\code{treelabel} vector}

\item{new_tree}{an igraph object that describes the new tree or
a function that creates a tree from the old tree.}

\item{tree_root}{the name of the tree root. Defaults to the
to the name of the tree root of \code{x}.}

\item{...}{arguments passed to \code{\link{treelabel}}}

\item{.p}{predicate that is used to filter the vertices of the tree. Either a character vector or
a function that receives the names of the vertices and indicates the selection which ones to keep.}

\item{new_root}{the name of a vertex in the tree that will be
used as the root of the cut tree.}
}
\value{
a \code{treelabel} vector described by the new tree.
}
\description{
Modify tree of an existing treelabel
}
