% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_abundance_change.R
\name{vars}
\alias{vars}
\title{Make \code{vars} available}
\usage{
vars(...)
}
\arguments{
\item{...}{Variables or expression are automatically quoted.

See \code{ggplot2::vars} or \code{rlang::quos} for more information}
}
\description{
Make \code{vars} available
}
