% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelabel.R
\name{treelabel}
\alias{treelabel}
\alias{treelabel.matrix}
\alias{treelabel.data.frame}
\alias{treelabel.list}
\alias{treelabel.numeric}
\alias{treelabel.logical}
\alias{treelabel.character}
\alias{treelabel.factor}
\alias{treelabel.missing}
\title{Make a tree label vector}
\usage{
treelabel(x, tree, tree_root = "root", ...)

\method{treelabel}{matrix}(x, tree, tree_root = "root", ...)

\method{treelabel}{data.frame}(
  x,
  tree,
  tree_root = "root",
  id = "id",
  label = "label",
  score = "score",
  ...
)

\method{treelabel}{list}(x, tree, tree_root = "root", ...)

\method{treelabel}{numeric}(x, tree, tree_root = "root", ...)

\method{treelabel}{logical}(x, tree, tree_root = "root", ...)

\method{treelabel}{character}(x, tree, tree_root = "root", ...)

\method{treelabel}{factor}(x, tree, tree_root = "root", ...)

\method{treelabel}{missing}(x, tree, tree_root = "root", ...)
}
\arguments{
\item{x}{object that is converted to a \code{treelabel}}

\item{tree}{\code{igraph} object that represents the hierarchical relationship
of the labels. Must satisfy \code{igraph::is_tree}. (_)}

\item{tree_root}{the name of the root of the tree. Default: \code{"root"}.}

\item{...}{additional arguments. Provided for compatability.}

\item{id, label, score}{strings that specify the names of the respective
columns in the \code{treelabel.data.frame} constructor.}
}
\description{
Make a tree label vector
}
