% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelabel.R
\name{treelabel}
\alias{treelabel}
\alias{treelabel.data.frame}
\alias{treelabel.list}
\alias{treelabel.numeric}
\alias{treelabel.character}
\alias{treelabel.factor}
\title{Make a tree label vector}
\usage{
treelabel(x, tree, ...)

\method{treelabel}{data.frame}(x, tree, id = "id", label = "label", score = "score")

\method{treelabel}{list}(x, tree, ...)

\method{treelabel}{numeric}(x, tree, ...)

\method{treelabel}{character}(x, tree, ...)

\method{treelabel}{factor}(x, tree, ...)
}
\description{
Make a tree label vector
}
