% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelabel_from_dataframe.R
\name{treelabel_from_dataframe}
\alias{treelabel_from_dataframe}
\title{Convert data.frame to treelabel}
\usage{
treelabel_from_dataframe(
  x,
  tree,
  tree_root = "root",
  id = "id",
  label = "label",
  score = "score",
  name = "treelabel",
  propagate_up = c("sum", "cumsum", "none"),
  ...
)
}
\arguments{
\item{id, label, score}{strings that specify the names of the respective
columns in the \code{treelabel.data.frame} constructor.}

\item{label, score}{the column names of \code{x} that specify the names and confidence scores
of each entry}

\item{name}{the name of the \code{treelabel} column in the output \code{data.frame}.}

\item{id}{the column name of \code{x} specifying which elements belong together.}
}
\value{
a \code{data.frame} with one row per unique \code{id} and two columns:
The first is a vector of the ids and the second is the \code{treelabel} vector.
}
\description{
Convert data.frame to treelabel
}
